requirements.txt
cycler==0.10.0
glcontext==2.3.2
kiwisolver==1.3.1
matplotlib==3.3.4
moderngl==5.6.3
moderngl-window==2.3.0
multipledispatch==0.6.0
numpy==1.20.1
onnxruntime-gpu==1.6.0
opencv-python==4.5.1.48
Pillow==8.1.0
protobuf==3.14.0
pyglet==1.5.15
pyparsing==2.4.7
pyrr==0.10.3
python-dateutil==2.8.1
scipy==1.6.0
six==1.15.0
